net: qualcomm: rmnet: Avoid *_hdr() helpers while adding linear GSO
These various helpers rely on the network_offset and transport_offset
fields in the SKB being set to appropriate values, and they might not be
set appropriately in all cases (e.g. reusing the coalesced SKB from the
lower levels). Access the headers for updating the GSO information
directly via skb->data to ensure the correct bytes are updated.

Change-Id: I4ac2b3118c190587d8178b1d780c9ec842438c86
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>